requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sentence-transformers
faiss-cpu
PyPDF2
openai
slowapi
groq
